Conform with expected input format
