+++ packages/loopring_v3/test/testDebugTools.ts
@@ -1,5 +1,6 @@
 import BN = require("bn.js");
 import fs = require("fs");
+import { AmmPool } from "./ammUtils";
 import { ExchangeTestUtil, OnchainBlock } from "./testExchangeUtil";
 import { BlockCallback } from "./types";
 
@@ -21,7 +22,7 @@ contract("Exchange", (accounts: string[]) => {
     it.skip("submitBlocks tx data", async () => {
       const blockDirectory = "./blocks/";
 
-      const blockNames = ["block_2_3", "block_2_4", "block_2_5"];
+      const blockNames = ["block_2_3", "block_2_4"];
       const outputFilename = "./blocks/result.json";
 
       const onchainBlocks: OnchainBlock[] = [];
@@ -29,16 +30,21 @@ contract("Exchange", (accounts: string[]) => {
       for (const blockName of blockNames) {
         const baseFilename = blockDirectory + blockName;
         const auxDataFilename = baseFilename + "_auxiliaryData.json";
-        const callbacksFilename = baseFilename + "_callbacks.json";
+        //const callbacksFilename = baseFilename + "_callbacks.json";
         const proofFilename = baseFilename + "_proof.json";
+        const blockInfoFilename = baseFilename + "_info.json";
         const blockFilename = baseFilename + ".json";
 
         const auxiliaryData = JSON.parse(
           fs.readFileSync(auxDataFilename, "ascii")
         );
-        const callbacks = JSON.parse(
+        /*const callbacksReference = JSON.parse(
           fs.readFileSync(callbacksFilename, "ascii")
+        );*/
+        const blockInfo = JSON.parse(
+          fs.readFileSync(blockInfoFilename, "ascii")
         );
+        //console.log(callbacksReference);
         const proof = ctx.readProof(proofFilename);
 
         // Read in the block
@@ -54,6 +60,13 @@ contract("Exchange", (accounts: string[]) => {
           proof
         );
         console.log(onchainBlock);
+
+        // Read the AMM transactions
+        const callbacks: BlockCallback[] = [];
+        for (const ammTx of blockInfo.ammTransactions) {
+          callbacks.push(AmmPool.getBlockCallback(ammTx));
+        }
+        //console.log(callbacks);
 
         onchainBlocks.push(onchainBlock);
         blockCallbacks.push(callbacks);
